refactor(ppt): split character-style resolution into ppt_style
Mirror the doc_style/xls_style split that landed on main: move the
StyleTextPropAtom parsing (TextCFRun, parse_style_text_prop_atom) out of
ppt_io — which goes back to pure istream record/text primitives — and
resolve_style/StyleContext plus the 18pt default (now
default_character_style, matching doc) out of ppt_parser.
