+++ CMakeLists.txt
@@ -163,6 +163,7 @@ set(ODR_SOURCE_FILES
         "src/odr/internal/oldms/presentation/ppt_element_registry.cpp"
         "src/odr/internal/oldms/presentation/ppt_io.cpp"
         "src/odr/internal/oldms/presentation/ppt_parser.cpp"
+        "src/odr/internal/oldms/presentation/ppt_style.cpp"
         "src/odr/internal/oldms/oldms_file.cpp"
 
         "src/odr/internal/ooxml/presentation/ooxml_presentation_document.cpp"
