+++ src/odr/internal/oldms/presentation/ppt_io.hpp
@@ -1,15 +1,11 @@
 #pragma once
 
-#include <odr/style.hpp>
-
 #include <odr/internal/oldms/presentation/ppt_structs.hpp>
 
 #include <cstdint>
 #include <iosfwd>
-#include <optional>
 #include <string>
 #include <string_view>
-#include <vector>
 
 namespace odr::internal::oldms::presentation {
 
@@ -45,24 +41,5 @@ std::string read_raw_text_bytes(std::istream &in, std::uint32_t rec_len);
 /// Decodes TextBytesAtom characters (each byte a code point 0x00-0xFF) to
 /// UTF-8.
 std::string decode_text_bytes(std::string_view bytes);
-
-/// One character-formatting run of a StyleTextPropAtom ([MS-PPT] 2.9.46);
-/// each field is set only when its CFMasks bit is.
-struct TextCFRun final {
-  std::uint32_t count{0}; //< characters this run covers
-  std::optional<bool> bold;
-  std::optional<bool> italic;
-  std::optional<bool> underline;
-  std::optional<std::uint16_t> font_ref; //< index into the FontCollection
-  std::optional<std::int16_t> font_size; //< points
-  std::optional<Color> color; //< explicit sRGB only; scheme indexes are unset
-};
-
-/// Parses a StyleTextPropAtom body ([MS-PPT] 2.9.44): the paragraph-level runs
-/// are skipped, the character-level runs returned. `char_count` is the
-/// corresponding text's length in characters plus one (the runs also cover an
-/// implicit final paragraph mark); fewer covered characters end the parse.
-std::vector<TextCFRun> parse_style_text_prop_atom(std::string_view body,
-                                                  std::size_t char_count);
 
 } // namespace odr::internal::oldms::presentation
